# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "自托管通知发送服务 API（邮箱/企微/钉钉/飞书/PushPlus 多渠道投递）。",
        "title": "Notice Service API",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/api/webhook/{api_key}": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Webhook"
                ],
                "summary": "用 API Key 触发任务（无需登录）",
                "parameters": [
                    {
                        "type": "string",
                        "description": "任务 API Key",
                        "name": "api_key",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "变量",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}